GH-3: Use import modules feature from ES6 Valid only for the next browsers: https://caniuse.com/#search=import
this can be done for all the browsers using tools like webpack
