+++ src/index.html
@@ -11,6 +11,5 @@
 <body>
   <h1>Spring challenge</h1>
   <div id="user-posts"></div>
-  <script src="index.js"></script>
-</body>
+  <script type="module" src="index.js"></script>
 </html>
