+++ src/utils/get.js
@@ -1,0 +1,3 @@
+const get = url => fetch(url).then(response => response.json());
+
+export default get;
